+++ src/main/java/com/github/tth05/jindex/IndexedField.java
@@ -14,7 +14,7 @@ public class IndexedField {
 
     public native String getName();
 
-    public native short getAccessFlags();
+    public native int getAccessFlags();
 
     public native String getDescriptorString();
     public native String getGenericSignatureString();
